requirements.txt
Django==5.2.7
requests>=2.31,<3
Pillow>=10.0,<11
python-dotenv>=1.0.0
djangorestframework>=3.16.0
django-cors-headers>=4.0.0
channels>=4.0.0
channels-redis>=4.0.0
daphne>=4.0.0

# Production dependencies for Render
psycopg2-binary>=2.9.9  # PostgreSQL adapter
gunicorn>=21.2.0  # WSGI server
whitenoise>=6.6.0  # Static files serving
dj-database-url>=2.1.0  # Database URL parsing
